requirements.txt
hyperon== 0.2.3 
Flask == 3.0.3
Werkzeug==3.0.2
biocypher>=0.5.4
python-dotenv==1.0.1
PyYAML==6.0.1
flask-cors == 4.0.1
Flask-Mail ==0.10.0
openpyxl == 3.1.5
PyJWT == 2.9.0
pymongoose == 1.3.8
openai == 1.51.2
pytest == 8.3.3
gunicorn == 23.0.0
Flask-Limiter == 3.8.0 
tiktoken==0.8.0
httpx==0.27.2
networkx==3.4.2
Flask-SocketIO==5.5.1
flask-redis==0.4.0
nanoid==2.0.0
networkx==3.4.2
elasticsearch==8.11.1
neo4j==5.16.0